Added navigation stack manager
This relates to #9

<me@themikkel.dk>
