+++ src/screens/App.tsx
@@ -1,27 +1,27 @@
 import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, View } from 'react-native';
+import { StyleSheet, View, Button } from 'react-native';
 import { Text } from '../components/atoms/Text';
 import { CarsProvider } from '../components/molecules/CarsContext';
 import { LocationsProvider } from '../components/molecules/LocationContext';
+import { NavigationContainer } from '@react-navigation/native';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import Home from './Home';
+import { CarBrowser } from './CarBrowser';
+
+const stack = createNativeStackNavigator();
 
 export default function App() {
   return (
     <CarsProvider>
       <LocationsProvider>
-        <View style={styles.container}>
-          <Text>Hello world!</Text>
-          <StatusBar style="auto" />
-        </View>
+        <NavigationContainer>
+          <stack.Navigator initialRouteName="Home">
+            <stack.Screen name="Home" component={Home} />
+            <stack.Screen name="CarBrowser" component={CarBrowser} />
+          </stack.Navigator>
+        </NavigationContainer>
       </LocationsProvider>
     </CarsProvider>
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
